ポップアップ画面にカレンダーと設定を表示する
