+++ src/theme/theme.ts
@@ -1,9 +1,10 @@
 import { createMuiTheme } from '@material-ui/core';
+import { blue, amber } from '@material-ui/core/colors';
 
 export const theme = createMuiTheme({
     palette: {
         type: 'light',
-        primary: { main: '#2962ff' },
-        secondary: { main: '#ffab00' },
+        primary: { main: `${blue[600]}` },
+        secondary: { main: `${amber[700]}` },
     },
 });
